src/posts/entities/post.repository.ts
import { EntityRepository, Repository } from 'typeorm';
import { PostCreationDto } from '../dto/postCreation.dto';
import { Post } from './post.entity';

@EntityRepository(Post)
export class PostRepository extends Repository<Post> {
  public async createPost(postCreationDto: PostCreationDto): Promise<Post> {
    const { caption, type, is_hidden } = postCreationDto;
    const post = this.create();
    post.caption = caption;
    post.type = type;
    post.is_hidden = is_hidden;
    post.user_id = 1;
    post.created = false;
    post.ready = false;
    return await this.save(post);
  }
  async getAllPosts() {
    const posts = await this.find({ relations: ['groups', 'groups.options'] });
    const posts_count = posts.length;
    const response = { postsCount: posts_count, posts: posts };
    return response;
  }
}
